Add files via upload
I’ve revised the demo images, with no content removed—just adjustments to the active images. All of the images are now consistently sized at 500x500, which ensures a nice, uniform appearance when clicked in the modal.

If we’re going to pursue this project, I think it’s important we do it the right way.