Refactor CSS to standardize margin and padding units for .welcome_note
Margin and Padding Units of 0 (zero) should not have identifiers. e.g., px or rem.

Also, I consolidated the rules for .welcome_note.  The first rule had four padding listings that could be combined into one.

The second rule had a redundant vertical-align as the vertical-align: baseline would be the observed item.

I removed the final .welcome_note rule, as coming after the others in the file, it would invalidate the earlier settings for padding.

NOTE: There is no longer a requirement for the semi-colon (;) at the end of the last item in a rule, but in a file where inserts are often made, it might be a good idea to standardize getting one last (;) in before the (}).